+++ license-header.js
@@ -3,8 +3,3 @@
  *
  * For license information see LICENSE file.
  */
-
-export interface EnumItemDescription<T> {
-  value: T;
-  description: string;
-}