+++ projects/abraxas/voting-lib/src/lib/components/sidebar/sidebar.component.ts
@@ -1,6 +1,7 @@
-/*!
- * (c) Copyright 2022 by Abraxas Informatik AG
- * For license information see LICENSE file
+/**
+ * (c) Copyright 2024 by Abraxas Informatik AG
+ *
+ * For license information see LICENSE file.
  */
 
 import { animate, state, style, transition, trigger } from '@angular/animations';