+++ projects/abraxas/voting-lib/src/lib/services/rxjs.utils.ts
@@ -1,37 +1,37 @@
-/*!
- * (c) Copyright 2022 by Abraxas Informatik AG
- * For license information see LICENSE file
+/**
+ * (c) Copyright 2024 by Abraxas Informatik AG
+ *
+ * For license information see LICENSE file.
  */
 
 import { defer, Observable, OperatorFunction, timer } from 'rxjs';
-import { concatMap, retryWhen, tap } from 'rxjs/operators';
+import { retry, tap } from 'rxjs/operators';
 
 const waitTime = 200;
 const maxWaitTime = 30000; // 30 seconds
 
-function getWaitTime(attempt: number): number {
-  return Math.min(waitTime * Math.pow(2, attempt), maxWaitTime);
-}
-
-export function retryForeverWithBackoff<T>(): OperatorFunction<T, T> {
+export function retryForeverWithBackoff<T>(onRetry?: () => {}): OperatorFunction<T, T> {
   return (src: Observable<T>) =>
     defer(() => {
-      let attempts = 0;
-
       // retry on failure after timeout
       return src.pipe(
-        retryWhen(errs =>
-          errs.pipe(
-            concatMap(() => {
-              attempts++;
-              const retryIn = getWaitTime(attempts);
-              return timer(retryIn);
-            }),
-          ),
-        ),
-
-        // reset attempts on successful emit
-        tap(() => (attempts = 0)),
+        retry({
+          delay: (error, retryCount) => {
+            const retryIn = getWaitTime(retryCount);
+            return timer(retryIn).pipe(
+              tap(() => {
+                if (onRetry) {
+                  onRetry();
+                }
+              }),
+            );
+          },
+          resetOnSuccess: true,
+        }),
       );
     });
+}
+
+function getWaitTime(attempt: number): number {
+  return Math.min(waitTime * Math.pow(2, attempt), maxWaitTime);
 }